requirements.txt
alembic==1.13.1
amqp==5.2.0
billiard==4.2.0
blinker==1.7.0
boto3==1.34.32
botocore==1.34.32
cached-property==1.5.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.14
duckdb==0.9.2
duckdb_engine==0.10.0
Flask==3.0.1
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
limits==3.7.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.3
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydub==0.25.1
pyfiglet==1.0.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rich==13.7.0
s3transfer==0.10.0
sedona==1.0.4
six==1.16.0
SQLAlchemy==2.0.25
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
